test: 각 피스별 테스트를 [piece명]SpaceTest로 변경
실제 이동은 Space에서 수행하기 때문에, test 내용에 맞게 package 및 class명 변경
